refactor : 레시피 수정 API 리팩토링
리팩토링 내용
- 기존 API는 레시피 수정시 대표 이미지를 변경하지 않아도 전송해야 했지만 이제는 전송하지 않으면 이미지의 대해서는 수정하지 않도록 리팩토링

+++ src/main/java/com/team/RecipeRadar/domain/recipe/application/RecipeServiceImpl.java
@@ -15,6 +15,7 @@ import com.team.RecipeRadar.global.Image.domain.UploadFile;
 import com.team.RecipeRadar.global.aws.S3.application.S3UploadService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
@@ -154,11 +155,14 @@ public class RecipeServiceImpl implements RecipeService{
         Recipe recipe = recipeRepository.findById(recipeId).orElseThrow(() -> new NoSuchElementException("해당 레시피를 찾을수 없습니다."));
 
         UploadFile uploadFile = imgRepository.findByRecipe_Id(recipeId).get();
-        if(!uploadFile.getOriginFileName().equals(file.getOriginalFilename())){
-            s3UploadService.deleteFile(uploadFile.getStoreFileName());
-            String storedFile = s3UploadService.uploadFile(file);
-            uploadFile.update(storedFile,file.getOriginalFilename());
-            imgRepository.save(uploadFile);
+
+        if(file!=null) {
+            if (!uploadFile.getOriginFileName().equals(file.getOriginalFilename())) {
+                s3UploadService.deleteFile(uploadFile.getStoreFileName());
+                String storedFile = s3UploadService.uploadFile(file);
+                uploadFile.update(storedFile, file.getOriginalFilename());
+                imgRepository.save(uploadFile);
+            }
         }
 
         List<Map<String, String>> cookeSteps = recipeUpdateRequest.getCookeSteps();
